artik710: support zigbee firmware upgrade
This patch supports zigbee firmware upgrade for artik710.
The firmware file is located /sbin directory and check the
firmware version before starting the sd recovery stage.
After firmware fusing is done, the zigbee f/w version will
be also checked to judge whether the fw upgrade is correct.
If there is a thread firmware, the thread firmware will be
not flashed and version check will be skipped.

Change-Id: I09e2926164b41449cb7409ad55d9a1238baf1712
Signed-off-by: Chanho Park <chanho61.park@samsung.com>