artik530: support zigbee firmware upgrade
This patch supports zigbee firmware upgrade for artik530.
The firmware file is located /sbin directory and check the
firmware version before starting the sd recovery stage.
After firmware fusing is done, the zigbee f/w version will
be also checked to judge whether the fw upgrade is correct.
If there is a thread firmware, the thread firmware will be
not flashed and version check will be skipped.

Change-Id: If62af21f9e5a57293658125cdacdca1fe5238d33
Signed-off-by: Chanho Park <chanho61.park@samsung.com>
Signed-off-by: Joonghong Park <joonghong.park@samsung.com>